Update configuration of outbound/source dependency
The Link class now uses the 'context' field of the table factory config
to specify the source table class which is referenced in the table
definition of the outbound table class. This ensures that the foreign
key reference to the source table can be resolved when datajoint tries
to declare the outbound table.